Don't generate excessive hash tables during reads.
Most read calls don't need the "read objects" hash tables, so if we
don't use them, save them for the next read call, so as to reduce the
allocations done.

* src/lread.c (read_internal_start): After reading, reset the hash
table variables to nil only if something was added to the hash
tables.
(readevalloop): Likewise.  Initialize the variables only if they're
not already set to empty hash tables.