Don't get into an error loop if dumped.elc isn't found.
When temacs is run, if dumped.elc isn't in the right place (currently
$srctop/src/dumped.elc, though it's generated in the build tree's src
directory), starting in X11 mode will produce a message about not
finding it, but starting in tty mode will produce an error loop
because internal-echo-keystrokes-prefix isn't defined.

This patch traps the error and kills the Emacs session with an exit
code of 42, and no message.  This is worse than the X11 behavior, but
better than the tty behavior.

It's not a long term fix though.

* src/emacs.c (main): If dumped.elc isn't found, exit with an error
code.
(syms_of_emacs): Define Qcondition_case.